resolve #24, #22, #17. Started #27
